tests/wpt/web-platform-tests/wasm/wasm_local_iframe_test.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/load_wasm.js"></script>
<script>
  function runTests(iframe) {
      iframe = iframe.contentWindow;
      promise_test(async function() {
        var mod = await createWasmModule();
        assert_true(mod instanceof WebAssembly.Module);
        var ans = await new Promise((resolve, reject) => {
          iframe.postMessage(mod, '*');
          window.addEventListener("message", (reply) => resolve(reply.data), false);
        });
        assert_equals(ans, 43);
      }, "send wasm module to iframe");
  }
</script>
<iframe src="resources/frame.html" onload="runTests(this)"></iframe>
